pyproject.toml
[tool.poetry]
name = "mistral-pdf"
version = "0.1.0"
description = ""
authors = ["IsmaelMekene <ismaelmekene@gmail.com>"]
readme = "README.md"
packages = [{include = "mistral_pdf"}]

[tool.poetry.dependencies]
python = "^3.8"
farm-haystack = {extras = ["weaviate"], version = "^1.21.2"}
canals = "^0.8.1"
pymupdf = "^1.23.5"
pypdf2 = "^3.0.1"
nltk = "^3.8.1"
torch = "^2.1.0"
sentence-transformers = "^2.2.2"
llama-cpp-python = "^0.2.11"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"

[dependencies]
python = "^3.8"
cached-property = "^1.5.2"
fastapi = "^0.68.0"
dotenv = ">=0.0.5"
uvicorn = "^0.15.0"
python-dotenv = "^0.19.0"
python-multipart = "^0.0.5"
torch = "^1.8.1"
sentence-transformers = "^1.2.0"
haystack-ai = "^0.10.0"
canals = "^0.1.0"


[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.961"
types-python-dateutil = "^2.8.18"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[mypy]
python_version = "^3.8"
exclude = "tests/"

[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "-ra -q -vv --cov=mistral_pdf --cov-report=term"

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
omit = ["**/tests/test_*.py"]
fail_under = 50
